refactor: eatinorders.domain 패키지, order과 ordertable로 하위패키지 구분
