+++ src/main/java/kitchenpos/eatinorders/domain/order/EatInOrderRepository.java
@@ -1,8 +1,10 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.domain.order;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+
+import kitchenpos.eatinorders.domain.ordertable.OrderTable;
 
 public interface EatInOrderRepository {
     EatInOrder save(EatInOrder eatInOrder);
