+++ src/main/java/kitchenpos/eatinorders/domain/order/OrderLineItems.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.domain.order;
 
 import java.util.List;
