+++ src/main/java/kitchenpos/eatinorders/domain/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.domain.order;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, COMPLETED
