+++ src/main/java/kitchenpos/eatinorders/domain/ordertable/NumberOfGuests.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.domain.ordertable;
 
 import java.util.Objects;
