+++ src/main/java/kitchenpos/eatinorders/infra/JpaEatInOrderRepository.java
@@ -4,8 +4,8 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.eatinorders.domain.EatInOrder;
-import kitchenpos.eatinorders.domain.EatInOrderRepository;
+import kitchenpos.eatinorders.domain.order.EatInOrder;
+import kitchenpos.eatinorders.domain.order.EatInOrderRepository;
 
 public interface JpaEatInOrderRepository extends EatInOrderRepository, JpaRepository<EatInOrder, UUID> {
 }
