+++ src/test/java/kitchenpos/eatinorders/domain/NumberOfGuestsTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import kitchenpos.eatinorders.domain.ordertable.NumberOfGuests;
+
 class NumberOfGuestsTest {
 
     @ParameterizedTest
