+++ src/test/java/kitchenpos/eatinorders/domain/OrderLineItemTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import kitchenpos.eatinorders.domain.order.OrderLineItem;
+
 class OrderLineItemTest {
     @Test
     void menuId는_없을_수_없다() {
