+++ src/test/java/kitchenpos/eatinorders/domain/QuantityTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import kitchenpos.eatinorders.domain.order.Quantity;
+
 class QuantityTest {
     @ParameterizedTest
     @ValueSource(ints = {0, 1})
